\name{exprs}
\alias{exprs}
\docType{data}
\title{Gene Expression Dataset}
\description{
  Gene expression dataset, containing 100 genes and 14 samples(7 case and 7 control respectively).
}
\usage{data(exprs)}
\details{
In this dataset, there are 100 genes and 14 samples which consist of the expression dataset, in which 7 samples are in case groups and other 7 samples are in control groups.
}
\author{Mingli Lei }
\examples{
data(exprs)
}
\keyword{datasets}