\name{networkModule}
\alias{networkModule}
\docType{data}
\title{NetworkModule}
\description{
  Different gene sets or modules in the biological network or metabolic pathway.
}
\usage{data(networkModule)}
\details{
 In this networkModule, there are five modules consist of different genes.
}
\author{Mingli Lei, Jia Xu, Li-Ching Huang, Lily Wang, Jing Li }
\examples{
data(networkModule)
}
\keyword{datasets}